requirements.txt
glfw==2.10.0
numpy==2.4.4
Pillow==12.1.1
pyglm==2.8.3
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10